+++ feature/editor/src/test/kotlin/jo/codeide/feature/editor/ActionsFichiersEditorViewModelTest.kt
@@ -86,7 +86,7 @@ class ActionsFichiersEditorViewModelTest : BaseEditorViewModelTest() {
                     tooling,
                     TestDispatcherProvider(regleMain.dispatcher),
                 ),
-            horloge = horlogeOutil,
+            serviceGradle = serviceGradleTest,
             savedStateHandle = sauvegarde,
         )